src/requirements.txt
addict==2.4.0
asttokens==3.0.0
attrs==25.1.0
blinker==1.9.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cssselect2==0.7.0
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
defusedxml==0.7.1
depthai==2.29.0.0
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.0.3
fonttools==4.55.6
idna==3.10
importlib_metadata==8.6.1
ipython==8.31.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kids.cache==0.0.7
kiwisolver==1.4.8
libsass==0.23.0
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mock==5.1.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.2.2
open3d==0.19.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
perlin-noise==1.13
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.1
pyquaternion==0.9.9
PySide6==6.8.1.1
PySide6_Addons==6.8.1.1
PySide6_Essentials==6.8.1.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
QT-PyQt-PySide-Custom-Widgets==1.0.2
QtPy==2.4.2
qtsass==0.4.0
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
shiboken6==6.8.1.1
six==1.17.0
stack-data==0.6.3
tenacity==9.0.0
termcolor==2.5.0
threadpoolctl==3.5.0
tinycss2==1.4.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.6
widgetsnbextension==4.0.13
zipp==3.21.0
